+++ src/i_world.h
@@ -97,7 +97,7 @@ class IWorld {
     /// @param    type The type of feromone
     /// @param    value The value to set
     ///----------------------------------------------------------------------
-    virtual void set_feromone(PositionT p, FeromoneT type, int value) = 0;
+    virtual void set_feromone(PositionT p, FeromoneT type, float value) = 0;
 
     ///----------------------------------------------------------------------
     /// @brief      Get the strength of a specific feromone type at a given position
@@ -106,7 +106,7 @@ class IWorld {
     /// @return     feromone strength at the given position
     ///----------------------------------------------------------------------
 
-    [[nodiscard]] virtual int get_feromone_strength(FeromoneT type, PositionT pos) const = 0;
+    [[nodiscard]] virtual float get_feromone_strength(FeromoneT type, PositionT pos) const = 0;
 
     ///----------------------------------------------------------------------
     /// @brief      Get the temperature at a given position
